+++ components/src/lib/Modal/storybook/ConfirmModal/ConfirmModal.buttonPosition.stories.tsx
@@ -1,10 +1,10 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import PromptModal from '../PromptModal';
+import ConfirmModal from '../../ConfirmModal';
 
 const meta = {
-  title: 'PromptModal/buttonPosition',
-  component: PromptModal,
-} satisfies Meta<typeof PromptModal>;
+  title: 'ConfirmModal/buttonPosition',
+  component: ConfirmModal,
+} satisfies Meta<typeof ConfirmModal>;
 
 export default meta;
 
@@ -16,13 +16,11 @@ export const Row: Story = {
     onClose: () => {
       alert('모달 닫기');
     },
+    onConfirm: () => {
+      alert('확인')
+    },
     title: '제목입니다',
     description: '설명입니다',
-
-    onConfirm: (value: string) => {
-      alert(`${value} 전달됨`)
-    },
-    placeholder: 'placeholder',
   },
 };
 
@@ -33,12 +31,10 @@ export const Column: Story = {
     onClose: () => {
       alert('모달 닫기');
     },
+    onConfirm: () => {
+      alert('확인')
+    },
     title: '제목입니다',
     description: '설명입니다',
-
-    onConfirm: (value: string) => {
-      alert(`${value} 전달됨`)
-    },
-    placeholder: 'placeholder',
   },
 };
