+++ components/src/lib/Modal/storybook/ConfirmModal/ConfirmModal.modalPosition.stories.tsx
@@ -1,47 +1,40 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import AlertModal from '../AlertModal';
+import ConfirmModal from '../../ConfirmModal';
 
 const meta = {
-  title: 'AlertModal/size',
-  component: AlertModal,
-} satisfies Meta<typeof AlertModal>;
+  title: 'ConfirmModal/modalPosition',
+  component: ConfirmModal,
+} satisfies Meta<typeof ConfirmModal>;
 
 export default meta;
 
 type Story = StoryObj<typeof meta>;
 
-export const Small: Story = {
+export const Center: Story = {
   args: {
     isOpened: true,
     onClose: () => {
       alert('모달 닫기');
+    },
+    onConfirm: () => {
+      alert('확인')
     },
     title: '제목입니다',
     description: '설명입니다',
   },
 };
 
-export const Medium: Story = {
+export const Bottom: Story = {
   args: {
-    size: 'medium',
+    modalPosition: 'bottom',
     isOpened: true,
     onClose: () => {
       alert('모달 닫기');
+    },
+    onConfirm: () => {
+      alert('확인')
     },
     title: '제목입니다',
     description: '설명입니다',
   },
 };
-
-export const Large: Story = {
-  args: {
-    size: 'large',
-    isOpened: true,
-    onClose: () => {
-      alert('모달 닫기');
-    },
-    title: '제목입니다',
-    description: '설명입니다',
-  },
-};
-
