+++ components/src/lib/Modal/storybook/Modal/Modal.buttonPosition.stories.tsx
@@ -1,8 +1,8 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import Modal from '../Modal';
+import Modal from '../../Modal';
 
 const meta = {
-  title: 'Modal/size',
+  title: 'Modal/buttonPosition',
   component: Modal,
 } satisfies Meta<typeof Modal>;
 
@@ -10,7 +10,7 @@ export default meta;
 
 type Story = StoryObj<typeof meta>;
 
-export const Small: Story = {
+export const Center: Story = {
   args: {
     isOpened: true,
     onClose: () => {
@@ -33,33 +33,9 @@ export const Small: Story = {
   },
 };
 
-export const Medium: Story = {
+export const Bottom: Story = {
   args: {
-    size: 'medium',
-    isOpened: true,
-    onClose: () => {
-      alert('모달 닫기');
-    },
-    title: '제목입니다',
-    description: '설명입니다',
-    primaryButton: {
-      text: '확인',
-      onClick: () => {
-        alert('확인');
-      },
-    },
-    secondaryButton: {
-      text: '취소',
-      onClick: () => {
-        alert('취소');
-      },
-    },
-  },
-};
-
-export const Large: Story = {
-  args: {
-    size: 'large',
+    buttonPosition: 'column',
     isOpened: true,
     onClose: () => {
       alert('모달 닫기');
