+++ components/src/lib/Modal/storybook/PromptModal/PromptModal.buttonPosition.stories.tsx
@@ -1,8 +1,8 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import PromptModal from '../PromptModal';
+import PromptModal from '../../PromptModal';
 
 const meta = {
-  title: 'PromptModal/size',
+  title: 'PromptModal/buttonPosition',
   component: PromptModal,
 } satisfies Meta<typeof PromptModal>;
 
@@ -10,7 +10,7 @@ export default meta;
 
 type Story = StoryObj<typeof meta>;
 
-export const Small: Story = {
+export const Row: Story = {
   args: {
     isOpened: true,
     onClose: () => {
@@ -26,9 +26,9 @@ export const Small: Story = {
   },
 };
 
-export const Medium: Story = {
+export const Column: Story = {
   args: {
-    size: 'medium',
+    buttonPosition: 'column',
     isOpened: true,
     onClose: () => {
       alert('모달 닫기');
@@ -42,21 +42,3 @@ export const Medium: Story = {
     placeholder: 'placeholder',
   },
 };
-
-export const Large: Story = {
-  args: {
-    size: 'large',
-    isOpened: true,
-    onClose: () => {
-      alert('모달 닫기');
-    },
-    title: '제목입니다',
-    description: '설명입니다',
-
-    onConfirm: (value: string) => {
-      alert(`${value} 전달됨`)
-    },
-    placeholder: 'placeholder',
-  },
-};
-
